gh-pages.css: add some empty space below the #site-header
The styling of the other SE-EDU websites (e.g. AddressBook-Level4) is
such that headers will have some margins to separate them with
whitespace from the rest of the document.

This is not so with Bootstrap, where headers have extremely tiny
margins. This causes the document body to be uncomfortably close to the
site navigation bar, which looks bad.

Fix this by adding some space below the #site-header.