Link gh-pages.css stylesheet in <head> element
We currently link the gh-pages.css stylesheet, which contains our
website-specific style rules, in the <body> element.

This is not valid HTML.

Fix this by using Markbind's `head` feature to link the gh-pages.css
stylesheet in the <head> element.

This is done by introducing the _markbind/head/head.md document, and by
adding the following to the frontmatter of every markdown file:

    head: head.md

The following script is used to accomplish this:

    for x in *.md $(find contents -name '*.md'); do
        sed -i '/^  footer: footer\.md/a\  head: head.md' "$x"
    done