navbar: add navbar-expand class
The items of the navigation bar are displayed as a column, which looks
ugly and is not consistent with the rest of the SE-EDU website.

Fix this by adding the `navbar-expand` class, which is a facility
provided by Bootstrap to display the navigation bar items in a row at
all times.