gh-pages.css: make the site navbar stretch across the entire page
In the official SE-EDU design, the site navbar at the top should stretch
across the entire page.

However, for the Learning Resources site, this does not happen. The site
navbar is constrained to a maximum width of 1000px and is centered in
the page. This makes the design of the Learning Resources site
inconsistent with the rest of the SE-EDU sites.

This is because a recent change in MarkBind[1] wraps the entire page
(minus the footer) in a #content-wrapper div, and the #content-wrapper
div is styled as:

    #content-wrapper {
        margin: 0 auto;
        max-width: 1000px;
        overflow-x: hidden;
        padding: 0 20px;
        width: 100%;
    }

thus leading to the current behavior.

Fix this by overriding #content-wrapper's `max-width` and `padding`
properties, so that the site navbar will stretch acorss the entire page.

[1]: https://github.com/MarkBind/markbind/commit/c20e9412987eec32e8ff9cb57eaf4d09da0fed5e