Improve NodeJS chapter
The NodeJS has some language inconsistencies.
Some of the sections go deeply into usage details of nodeJS and npm as
well.
The 8th benefit, npm having a large ecosystem of modules, seems very
similar to the 7th benefit as well.

Let's correct these language inconsistencies.
Let's also trim some of the deep details of nodeJS that are close to
tutorial territory, or move some of these into popovers and tooltips to
make things more succinct.
Let's also add a diagram for the event loop benefit, which can help
with understanding its technical nature.
Let's also merge the 8th benefit with the 7th to shorten the article.